requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
rich==13.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# AI Integration
groq==0.4.1
openai==1.3.0

# Tool Dependencies (akan diinstall di VM)
# nmap, nikto, hydra, sqlmap, gobuster, dll
